clisqlexec: split the table formatting into separate files
Release note: None
