+++ pkg/cli/clisqlexec/format_table.go
@@ -11,21 +11,11 @@
 package clisqlexec
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"html"
 	"io"
 	"strings"
-	"text/tabwriter"
-	"time"
-	"unicode/utf8"
 
 	"github.com/cockroachdb/cockroach/pkg/cli/clisqlclient"
-	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
-	"github.com/cockroachdb/cockroach/pkg/util"
-	"github.com/cockroachdb/cockroach/pkg/util/encoding/csv"
-	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/errors"
 	"github.com/olekukonko/tablewriter"
 )
@@ -246,430 +236,6 @@ func render(
 		nRows++
 	}
 
-	return nil
-}
-
-type asciiTableReporter struct {
-	rows  int
-	table *tablewriter.Table
-	buf   bytes.Buffer
-	w     *tabwriter.Writer
-
-	tableBorderMode int
-}
-
-func newASCIITableReporter(tableBorderMode int) *asciiTableReporter {
-	n := &asciiTableReporter{tableBorderMode: tableBorderMode}
-	// 4-wide columns, 1 character minimum width.
-	n.w = tabwriter.NewWriter(&n.buf, 4, 0, 1, ' ', 0)
-	return n
-}
-
-func (p *asciiTableReporter) describe(w io.Writer, cols []string) error {
-	if len(cols) > 0 {
-		// Ensure that tabs are converted to spaces and newlines are
-		// doubled so they can be recognized in the output.
-		expandedCols := make([]string, len(cols))
-		for i, c := range cols {
-			p.buf.Reset()
-			fmt.Fprint(p.w, c)
-			_ = p.w.Flush()
-			expandedCols[i] = p.buf.String()
-		}
-
-		// Initialize tablewriter and set column names as the header row.
-		p.table = tablewriter.NewWriter(w)
-		p.table.SetAutoFormatHeaders(false)
-		p.table.SetAutoWrapText(false)
-		var outsideBorders, insideLines bool
-		// The following table border modes are taken from psql.
-		// https://www.postgresql.org/docs/12/app-psql.html
-		switch p.tableBorderMode {
-		case 0:
-			outsideBorders, insideLines = false, false
-		case 1:
-			outsideBorders, insideLines = false, true
-		case 2:
-			outsideBorders, insideLines = true, false
-		case 3:
-			outsideBorders, insideLines = true, true
-		}
-		p.table.SetBorder(outsideBorders)
-		p.table.SetRowLine(insideLines)
-		p.table.SetReflowDuringAutoWrap(false)
-		p.table.SetHeader(expandedCols)
-		p.table.SetTrimWhiteSpaceAtEOL(true)
-		// This width is sufficient to show a "standard text line width"
-		// on the screen when viewed as a single column on a 80-wide terminal.
-		//
-		// It's also wide enough for the output of SHOW CREATE on
-		// moderately long column definitions (e.g. including FK
-		// constraints).
-		p.table.SetColWidth(72)
-	}
-	return nil
-}
-
-func (p *asciiTableReporter) beforeFirstRow(w io.Writer, iter RowStrIter) error {
-	if p.table == nil {
-		return nil
-	}
-
-	p.table.SetColumnAlignment(iter.Align())
-	return nil
-}
-
-// asciiTableWarnRows is the number of rows at which a warning is
-// printed during buffering in memory.
-const asciiTableWarnRows = 10000
-
-func (p *asciiTableReporter) iter(_, ew io.Writer, _ int, row []string) error {
-	if p.table == nil {
-		return nil
-	}
-
-	if p.rows == asciiTableWarnRows {
-		fmt.Fprintf(ew,
-			"warning: buffering more than %d result rows in client "+
-				"- RAM usage growing, consider another formatter instead\n",
-			asciiTableWarnRows)
-	}
-
-	for i, r := range row {
-		p.buf.Reset()
-		fmt.Fprint(p.w, r)
-		_ = p.w.Flush()
-		row[i] = p.buf.String()
-	}
-	p.table.Append(row)
-	p.rows++
-	return nil
-}
-
-func (p *asciiTableReporter) doneRows(w io.Writer, seenRows int) error {
-	if p.table != nil {
-		p.table.Render()
-		p.table = nil
-	} else {
-		// A simple delimiter, like in psql.
-		fmt.Fprintln(w, "--")
-	}
-
-	fmt.Fprintf(w, "(%d row%s)\n", seenRows, util.Pluralize(int64(seenRows)))
-	return nil
-}
-
-func (p *asciiTableReporter) doneNoRows(_ io.Writer) error {
-	p.table = nil
-	return nil
-}
-
-type csvReporter struct {
-	mu struct {
-		syncutil.Mutex
-		csvWriter *csv.Writer
-	}
-	stop chan struct{}
-}
-
-// csvFlushInterval is the maximum time between flushes of the
-// buffered CSV/TSV data.
-const csvFlushInterval = 5 * time.Second
-
-func makeCSVReporter(w io.Writer, format TableDisplayFormat) (*csvReporter, func()) {
-	r := &csvReporter{}
-	r.mu.csvWriter = csv.NewWriter(w)
-	if format == TableDisplayTSV {
-		r.mu.csvWriter.Comma = '\t'
-	}
-
-	// Set up a flush daemon. This is useful when e.g. visualizing data
-	// from change feeds.
-	r.stop = make(chan struct{}, 1)
-	go func() {
-		ticker := time.NewTicker(csvFlushInterval)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ticker.C:
-				r.mu.Lock()
-				r.mu.csvWriter.Flush()
-				r.mu.Unlock()
-			case <-r.stop:
-				return
-			}
-		}
-	}()
-	cleanup := func() {
-		close(r.stop)
-	}
-	return r, cleanup
-}
-
-func (p *csvReporter) describe(w io.Writer, cols []string) error {
-	p.mu.Lock()
-	if len(cols) == 0 {
-		_ = p.mu.csvWriter.Write([]string{"# no columns"})
-	} else {
-		_ = p.mu.csvWriter.Write(cols)
-	}
-	p.mu.Unlock()
-	return nil
-}
-
-func (p *csvReporter) iter(_, _ io.Writer, _ int, row []string) error {
-	p.mu.Lock()
-	if len(row) == 0 {
-		_ = p.mu.csvWriter.Write([]string{"# empty"})
-	} else {
-		_ = p.mu.csvWriter.Write(row)
-	}
-	p.mu.Unlock()
-	return nil
-}
-
-func (p *csvReporter) beforeFirstRow(_ io.Writer, _ RowStrIter) error { return nil }
-func (p *csvReporter) doneNoRows(_ io.Writer) error                   { return nil }
-
-func (p *csvReporter) doneRows(w io.Writer, seenRows int) error {
-	p.mu.Lock()
-	p.mu.csvWriter.Flush()
-	p.mu.Unlock()
-	return nil
-}
-
-type rawReporter struct{}
-
-func (p *rawReporter) describe(w io.Writer, cols []string) error {
-	fmt.Fprintf(w, "# %d column%s\n", len(cols),
-		util.Pluralize(int64(len(cols))))
-	return nil
-}
-
-func (p *rawReporter) iter(w, _ io.Writer, rowIdx int, row []string) error {
-	fmt.Fprintf(w, "# row %d\n", rowIdx+1)
-	for _, r := range row {
-		fmt.Fprintf(w, "## %d\n%s\n", len(r), r)
-	}
-	return nil
-}
-
-func (p *rawReporter) beforeFirstRow(_ io.Writer, _ RowStrIter) error { return nil }
-func (p *rawReporter) doneNoRows(_ io.Writer) error                   { return nil }
-
-func (p *rawReporter) doneRows(w io.Writer, nRows int) error {
-	fmt.Fprintf(w, "# %d row%s\n", nRows, util.Pluralize(int64(nRows)))
-	return nil
-}
-
-type htmlReporter struct {
-	nCols int
-
-	escape   bool
-	rowStats bool
-}
-
-func (p *htmlReporter) describe(w io.Writer, cols []string) error {
-	p.nCols = len(cols)
-	fmt.Fprint(w, "<table>\n<thead><tr>")
-	if p.rowStats {
-		fmt.Fprint(w, "<th>row</th>")
-	}
-	for _, col := range cols {
-		if p.escape {
-			col = html.EscapeString(col)
-		}
-		fmt.Fprintf(w, "<th>%s</th>", strings.Replace(col, "\n", "<br/>", -1))
-	}
-	fmt.Fprintln(w, "</tr></thead>")
-	return nil
-}
-
-func (p *htmlReporter) beforeFirstRow(w io.Writer, _ RowStrIter) error {
-	fmt.Fprintln(w, "<tbody>")
-	return nil
-}
-
-func (p *htmlReporter) iter(w, _ io.Writer, rowIdx int, row []string) error {
-	fmt.Fprint(w, "<tr>")
-	if p.rowStats {
-		fmt.Fprintf(w, "<td>%d</td>", rowIdx+1)
-	}
-	for _, r := range row {
-		if p.escape {
-			r = html.EscapeString(r)
-		}
-		fmt.Fprintf(w, "<td>%s</td>", strings.Replace(r, "\n", "<br/>", -1))
-	}
-	fmt.Fprintln(w, "</tr>")
-	return nil
-}
-
-func (p *htmlReporter) doneNoRows(w io.Writer) error {
-	fmt.Fprintln(w, "</table>")
-	return nil
-}
-
-func (p *htmlReporter) doneRows(w io.Writer, nRows int) error {
-	fmt.Fprintln(w, "</tbody>")
-	if p.rowStats {
-		fmt.Fprintf(w, "<tfoot><tr><td colspan=%d>%d row%s</td></tr></tfoot>",
-			p.nCols+1, nRows, util.Pluralize(int64(nRows)))
-	}
-	fmt.Fprintln(w, "</table>")
-	return nil
-}
-
-type recordReporter struct {
-	cols        []string
-	colRest     [][]string
-	maxColWidth int
-}
-
-func (p *recordReporter) describe(w io.Writer, cols []string) error {
-	for _, col := range cols {
-		parts := strings.Split(col, "\n")
-		for _, part := range parts {
-			colLen := utf8.RuneCountInString(part)
-			if colLen > p.maxColWidth {
-				p.maxColWidth = colLen
-			}
-		}
-		p.cols = append(p.cols, parts[0])
-		p.colRest = append(p.colRest, parts[1:])
-	}
-	return nil
-}
-
-func (p *recordReporter) iter(w, _ io.Writer, rowIdx int, row []string) error {
-	if len(p.cols) == 0 {
-		// No record details; a summary will be printed at the end.
-		return nil
-	}
-	fmt.Fprintf(w, "-[ RECORD %d ]\n", rowIdx+1)
-	for j, r := range row {
-		lines := strings.Split(r, "\n")
-		for l, line := range lines {
-			colLabel := p.cols[j]
-			if l > 0 {
-				colLabel = ""
-			}
-			lineCont := "+"
-			if l == len(lines)-1 {
-				lineCont = ""
-			}
-			// Note: special characters, including a vertical bar, in
-			// the colLabel are not escaped here. This is in accordance
-			// with the same behavior in PostgreSQL. However there is
-			// special behavior for newlines.
-			contChar := " "
-			if len(p.colRest[j]) > 0 {
-				contChar = "+"
-			}
-			fmt.Fprintf(w, "%-*s%s| %s%s\n", p.maxColWidth, colLabel, contChar, line, lineCont)
-			for k, cont := range p.colRest[j] {
-				if k == len(p.colRest[j])-1 {
-					contChar = " "
-				}
-				fmt.Fprintf(w, "%-*s%s|\n", p.maxColWidth, cont, contChar)
-			}
-		}
-	}
-	return nil
-}
-
-func (p *recordReporter) doneRows(w io.Writer, seenRows int) error {
-	if len(p.cols) == 0 {
-		fmt.Fprintf(w, "(%d row%s)\n", seenRows, util.Pluralize(int64(seenRows)))
-	}
-	return nil
-}
-
-func (p *recordReporter) beforeFirstRow(_ io.Writer, _ RowStrIter) error { return nil }
-func (p *recordReporter) doneNoRows(_ io.Writer) error                   { return nil }
-
-type ndjsonReporter struct {
-	cols []string
-}
-
-func (n *ndjsonReporter) describe(w io.Writer, cols []string) error {
-	n.cols = cols
-	return nil
-}
-
-func (n *ndjsonReporter) beforeFirstRow(w io.Writer, allRows RowStrIter) error {
-	return nil
-}
-
-func (n *ndjsonReporter) iter(w, ew io.Writer, rowIdx int, row []string) error {
-	retMap := make(map[string]string, len(row))
-	for i := range row {
-		retMap[n.cols[i]] = row[i]
-	}
-	out, err := json.Marshal(retMap)
-	if err != nil {
-		return err
-	}
-	if _, err := ew.Write(out); err != nil {
-		return err
-	}
-	if _, err := ew.Write([]byte("\n")); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (n *ndjsonReporter) doneRows(w io.Writer, seenRows int) error {
-	return nil
-}
-
-func (n *ndjsonReporter) doneNoRows(w io.Writer) error {
-	return nil
-}
-
-type sqlReporter struct {
-	noColumns bool
-}
-
-func (p *sqlReporter) describe(w io.Writer, cols []string) error {
-	fmt.Fprint(w, "CREATE TABLE results (\n")
-	for i, col := range cols {
-		var colName bytes.Buffer
-		lexbase.EncodeRestrictedSQLIdent(&colName, col, lexbase.EncNoFlags)
-		fmt.Fprintf(w, "  %s STRING", colName.String())
-		if i < len(cols)-1 {
-			fmt.Fprint(w, ",")
-		}
-		fmt.Fprint(w, "\n")
-	}
-	fmt.Fprint(w, ");\n\n")
-	p.noColumns = (len(cols) == 0)
-	return nil
-}
-
-func (p *sqlReporter) iter(w, _ io.Writer, _ int, row []string) error {
-	if p.noColumns {
-		fmt.Fprintln(w, "INSERT INTO results(rowid) VALUES (DEFAULT);")
-		return nil
-	}
-
-	fmt.Fprint(w, "INSERT INTO results VALUES (")
-	for i, r := range row {
-		var buf bytes.Buffer
-		lexbase.EncodeSQLStringWithFlags(&buf, r, lexbase.EncNoFlags)
-		fmt.Fprint(w, buf.String())
-		if i < len(row)-1 {
-			fmt.Fprint(w, ", ")
-		}
-	}
-	fmt.Fprint(w, ");\n")
-	return nil
-}
-
-func (p *sqlReporter) beforeFirstRow(_ io.Writer, _ RowStrIter) error { return nil }
-func (p *sqlReporter) doneNoRows(_ io.Writer) error                   { return nil }
-func (p *sqlReporter) doneRows(w io.Writer, seenRows int) error {
-	fmt.Fprintf(w, "-- %d row%s\n", seenRows, util.Pluralize(int64(seenRows)))
 	return nil
 }
